build: switch toolchain, types, and linker to x86-64
Point the Makefile at the x86_64-elf cross toolchain (-m64, -mno-red-zone,
-mcmodel=kernel, elf64), add 64-bit types with a 64-bit size_t, retarget
the linker script to elf64-x86-64, and widen physical frame addresses to
uint64_t. Long-mode/GDT/IDT/ISR internals are handled separately.

+++ kernel/memory.h
@@ -6,8 +6,8 @@
 #define FRAME_SIZE 4096
 
 void memory_init(void);
-uint32_t alloc_frame(void);        // returns a free physical frame address, or 0 if none
-void free_frame(uint32_t addr);
+uint64_t alloc_frame(void);        // returns a free physical frame address, or 0 if none
+void free_frame(uint64_t addr);
 uint32_t frames_used(void);
 
 #endif
